Rename the DesignSystem folder to design-system for naming consistency
