+++ tailwind.config.js
@@ -5,7 +5,7 @@ module.exports = {
   content: [
     './src/pages/**/*.{js,ts,jsx,tsx}',
     './src/components/**/*.{js,ts,jsx,tsx}',
-    './src/DesignSystem/**/*.{js,ts,jsx,tsx}',
+    './src/design-system/**/*.{js,ts,jsx,tsx}',
   ],
   theme: {
     extend: {
